refactor: 점주, 손님 테이블 통합으로 회원 가입 기능 수정 - Member 엔티티에 userType 컬럼 추가 - SignUpParam -> SignUpDto로 변경 - MerchantRepository -> MemberRepository로 통합
