+++ src/main/java/fingerorder/webapp/entity/Member.java
@@ -27,6 +27,7 @@ public class Member{
     private String password;
     private String nickName;
     private String status;
+    private String userType;
     private LocalDateTime created_at;
     private LocalDateTime updated_at;
     private LocalDateTime deleted_at;
