+++ src/main/java/fingerorder/webapp/service/UserService.java
@@ -2,7 +2,7 @@ package fingerorder.webapp.service;
 
 import fingerorder.webapp.dto.UserDto;
 import fingerorder.webapp.parameter.SignInParam;
-import fingerorder.webapp.parameter.SignUpParam;
+import fingerorder.webapp.dto.SignUpDto;
 import fingerorder.webapp.parameter.UserEditParam;
 import fingerorder.webapp.parameter.UserParam;
 import java.util.List;
@@ -10,7 +10,7 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 
 
 public interface UserService extends UserDetailsService {
-	UserDto signUp(SignUpParam signUpParam);
+	UserDto signUp(SignUpDto signUpParam);
 	UserDto authenticate(SignInParam signInParam);
 	UserDto getUserInfo(UserParam userParam);
 	List<String> getRoles(SignInParam signInParam);
